requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
dotenv-python==0.0.1
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
rsa==4.9
Telethon==1.34.0
typing_extensions==4.11.0
yarl==1.9.4

pymongo~=4.7.2
fluent~=0.10.0